Module H1_parser

type bigstring = (char, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

type error =

| Msg of string

| Partial

val parse_headers : ?⁠pos:int -> ?⁠len:int -> bigstring -> (Cohttp.Header.t * int, error) Stdlib.result

val parse_request : ?⁠pos:int -> ?⁠len:int -> bigstring -> (Cohttp.Request.t * int, error) Stdlib.result

Attempts to parse a buffer into a HTTP request. If successful, it returns the parsed request and an offset value that indicates the starting point of unconsumed content left in the buffer.

val parse_chunk_length : ?⁠pos:int -> ?⁠len:int -> bigstring -> (int64 * int, error) Stdlib.result

val parse_chunk : ?⁠pos:int -> ?⁠len:int -> bigstring -> (string option * int, error) Stdlib.result